refactor(tests): split consolidated worktree.test.cjs along cleanup seam (≤ 800 LOC/file)
